29.01.2020

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
@@ -37,13 +38,18 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        $request->validate([
+           'name' => 'required|unique:categories,name',
+           'description' => 'required',
+        ]);
 
         $category = new Category;
         $category->name = $request->name;
         $category->description = $request->description;
+        $category->user_id = Auth::user()->id;
         $category->save();
 
-        return redirect()->route('categories.index');
+        return redirect()->route('categories.index')->with('success', 'Category added');
     }
 
     /**
@@ -77,22 +83,32 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        $request->validate([
+            'name' => 'required|unique:categories,name,'.$category->id,
+            'description' => 'required',
+        ]);
+
         $category->name = $request->name;
         $category->description = $request->description;
         $category->save();
 
-        return redirect()->route('categories.index');
+        return redirect()->route('categories.index')->with('success', 'Category updated');
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Category  $category
+     * @param \App\Category $category
      * @return \Illuminate\Http\Response
+     * @throws \Exception
      */
     public function destroy(Category $category)
     {
+        if($category->blogs->count() > 0){
+            return back()->with('danger', 'This category has linked blogs, cannot be deleted');
+        }
+
         $category->delete();
-        return redirect()->route('categories.index');
+        return redirect()->route('categories.index')->with('success', 'This category has deleted');
     }
 }
